pyproject.toml
[tool.poetry]
name = "maestro"
version = "0.1.0"
description = ""
authors = ["Łukasz Tshipenchko <dev@zxc.sx>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.3.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
paramiko = "^3.4.0"
matplotlib = "^3.8.2"
pillow = "^10.2.0"
text-to-image = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
